requirements.txt
mysql-connector-python
pydantic
pydantic[email]
pytest 
Faker